+++ NEWS.md
@@ -13,6 +13,8 @@
   a character or a factor and `label_eurostat_tables(x)` does not accept other
   than a character or a factor.
 * Cache files are now longer compressed.
+* For `get_eurostat()` a new argument `stringsAsFactors` to control the
+  factor conversion of variables.
 
 # eurostat 1.0.16